refactor(pages/blog/ posts/): switch next-mdx-enhenced to next-mdx-remote
