+++ package.json
@@ -11,16 +11,14 @@
     "preinstall": "npx only-allow pnpm"
   },
   "dependencies": {
-    "@mdx-js/loader": "^1.6.0",
     "babel-plugin-import-glob-array": "^0.2.0",
     "clean-webpack-plugin": "^3.0.0",
     "globby": "^11.0.2",
     "next": "11.1.4",
-    "next-mdx-enhanced": "^5.0.0",
+    "next-mdx-remote": "^4.4.1",
     "next-seo": "^6.5.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "reading-time": "^1.2.0",
     "rehype-highlight": "^4.0.0",
     "remark-autolink-headings": "^6.0.1",
     "remark-slug": "^6.0.0",
