+++ pages/_document.tsx
@@ -1,5 +1,4 @@
 import Document, { Html, Head, Main, NextScript, DocumentContext } from 'next/document';
-import React from 'react';
 
 export default class MyDocument extends Document {
   static async getInitialProps(ctx: DocumentContext) {
