+++ pages/about.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { NextSeo } from 'next-seo';
 import { HOST } from '../utils/constant';
