+++ pages/index.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export default function Home() {
   return (
     <>
