+++ posts/event-loop-of-browser.mdx
@@ -1,9 +1,9 @@
 ---
-layout: post
 title: 浏览器的事件循环(Event Loop)
 date: 2017-08-11
 description: 从HTML标准出发研究事件循环机制
 cover: /images/posts/event-loop-of-browser/cover.jpg
+slug: event-loop-of-browser
 ---
 
 #### 本文围绕浏览器的事件循环，而node.js有自己的另一套事件循环机制，不在本文讨论范围。网上的许多相关技术文章提到了`process.nextTick`和`setImmediate`两个node.js的API，这里不予讨论。
