+++ public/offline.html
@@ -8,6 +8,6 @@
     <title>Document</title>
   </head>
   <body>
-  You are already offline
+  You are offline
   </body>
 </html>
